PLANNER-401: Removed PureSingleStatistics and EffectiveStatisticMap from SingleBenchmarkResult
